refacto: moved fake npm_modules `get`, `my` and `hypem` to app/lib/ (#80) => simplified gitignore and Dockerfile
