+++ whydJS/app/lib/my/img/node-magick.js
